refactor: changes dir design pattern
