+++ src/routes/meal-routes.ts
@@ -1,9 +1,9 @@
-import { MealsController } from "@/adapters/controllers/meals-controller";
-import { MethodsMealsRepository } from "@/adapters/repositories/db-meals-repository";
 import { CreateMealUseCase } from "@/core/use-cases/meals/create-meal";
+import { MealsController } from "@/http/controllers/meals-controller";
+import { MethodsMealsRepository } from "@/repositories/db-meals-repository";
 import type { FastifyInstance } from "fastify";
 
-export async function   mealsRoutes(app: FastifyInstance) {
+export async function mealsRoutes(app: FastifyInstance) {
   const mealsRepository = new MethodsMealsRepository();
   const createMealUseCase = new CreateMealUseCase(mealsRepository);
   const mealsController = new MealsController(createMealUseCase);
